Move VibrationController to hardware package
