+++ app/src/main/java/io/github/vibrouter/managers/MainService.java
@@ -35,7 +35,7 @@ import com.google.gson.GsonBuilder;
 import java.util.Collections;
 import java.util.List;
 
-import io.github.vibrouter.VibrationController;
+import io.github.vibrouter.hardware.VibrationController;
 import io.github.vibrouter.models.DirectionsApiResult;
 import io.github.vibrouter.utils.GpsUtil;
